--- tmp-main/run.sh
@@ -1,25 +1,0 @@
-#!/bin/bash
-
-helpFunction()
-{
-   echo ""
-   echo "\t-t count of threads "
-   exit 1 # Exit script after printing help
-}
-
-while getopts "t:" opt
-do
-   case "$opt" in
-      t ) ThreadsCount="$OPTARG" ;;
-      ? ) helpFunction ;; # Print helpFunction in case parameter is non-existent
-   esac
-done
-
-# Print helpFunction in case parameters are empty
-if [ -z "$ThreadsCount" ] #|| [ -z "$parameterB" ] || [ -z "$parameterC" ]
-then
-   echo "Some or all of the parameters are empty";
-   helpFunction
-fi
-
-python3 test_synt.py $ThreadsCount
